RD-1: Events are not blocking, event states are.
Event states are blocking and events are always confirmed. When overwriting an event in the event store we check if the event is about to overwrite a blocking state at which point we retrieve the relevant event that caused that blocking state and direct the user to edit that event.